restructuring Plot.tsx file: created hook for MouseEvents
