+++ packages/suite-base/src/panels/Plot/Plot.style.ts
@@ -1,0 +1,34 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+import { buttonClasses } from "@mui/material";
+import { makeStyles } from "tss-react/mui";
+
+export const useStyles = makeStyles()((theme) => ({
+  tooltip: {
+    maxWidth: "none",
+  },
+  resetZoomButton: {
+    pointerEvents: "none",
+    position: "absolute",
+    display: "flex",
+    justifyContent: "flex-end",
+    paddingInline: theme.spacing(1),
+    right: 0,
+    left: 0,
+    bottom: 0,
+    width: "100%",
+    paddingBottom: theme.spacing(4),
+
+    [`.${buttonClasses.root}`]: {
+      pointerEvents: "auto",
+    },
+  },
+  canvasDiv: { width: "100%", height: "100%", overflow: "hidden", cursor: "crosshair" },
+  verticalBarWrapper: {
+    width: "100%",
+    height: "100%",
+    overflow: "hidden",
+    position: "relative",
+  },
+}));
